refactor: extract useChainId
