+++ packages/token-bridge-sdk/src/hooks/useChainId.ts
@@ -1,0 +1,16 @@
+import { useState, useEffect } from 'react'
+import { providers } from 'ethers'
+
+export function useChainId({ provider }: { provider: providers.Provider }) {
+  const [chainId, setChainId] = useState<number | undefined>(undefined)
+
+  useEffect(() => {
+    async function updateChainId() {
+      setChainId((await provider.getNetwork()).chainId)
+    }
+
+    updateChainId()
+  }, [provider])
+
+  return chainId
+}
